+++ CHANGELOG.md
@@ -14,10 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   - Update Shadow plugin to `7.0.0` version
   - Update Kotlin to `1.5.21` version
   - Set default KtLint version to `0.42.1`
+  - Rethink format task approach ([issue: #306](https://github.com/JLLeitschuh/ktlint-gradle/issues/306))
 
 ### Fixed
   - Pre-commit hook causing conflicts ([issue: #443](https://github.com/JLLeitschuh/ktlint-gradle/issues/443)) ([#502](https://github.com/JLLeitschuh/ktlint-gradle/pull/502))
-  - `ktlintFormat` create empty directories in `src/` dir ([issue: $423](https://github.com/JLLeitschuh/ktlint-gradle/issues/423))
+  - `ktlintFormat` create empty directories in `src/` dir ([issue: #423](https://github.com/JLLeitschuh/ktlint-gradle/issues/423))
 
 ### Removed
   - ?